New `Environment` field in LRP spec
This is a slice of `corev1.EnvVar`s so people can set env vars however
k8s supports it. In particular, the secret key ref is supported so that
secret values need not be exposed in the spec.

This also deprecates the LRP `Env` field. `Environment` supports the
plain text env vars that `Env` allows so `Env` can eventually be
removed.

Issue: https://github.com/cloudfoundry/korifi/issues/917