Apply task CPUMillis to job spec
CPUWeight was not being used in production code. It also had a strange
datatype of uint8 which was not suitable for a CPU request. Given that
korifi is eirini-controller's only client, we can remove this field
without problems.

Add a new field CPUMillis accurately describing the property. This is
applied to the request CPU in the job's container spec.

Issue: https://github.com/cloudfoundry/korifi/issues/1203